feat: add validation to Limit component
The data-flow on input is rearranged to better separate the string
entered by the user on one hand and the number stored in the store on the
other hand.

Unfortunately, when showing the validation error, then the height of the
TextInput changes and so the center-aligned items are rearranged. While
align-items: baseline would seem like the appropriate solution, it
doesn't seem to work with input elements. So, the only solution seems to
be to align the checkbox manually by giving it a extra bit of margin.

If the user is entering something invalid, then we show the error and
prevent the query from running, as requested by UX.

If the user merely deletes the number input, then we treat it as 100 and
reset it to that default value on Run Query.

The error message changes the content width of the TextInput component
and so its overall width adjusts. We need to define an explicit width to
prevent that.

Bug: T271943
Bug: T270179
Change-Id: Ia54d3ec0d1035f0630ab37d78516c601121c207a