Create action to handle value from datatype time
What remains to do here is implementing that the validation picks up on
this when it runs after the user clicked "Run Query". However, that code
was also touched by the Quantity query patches and those are not merged
yet. So that functionality will be implemented in an extra patch.

Note that we are using errorformat: 'plaintext', and formatversion: '2',
and both are affecting the structure of the errors that we are receiving
back from the API.

Bug: T277848
Change-Id: I5657d7f2065c74dde72bf8c6d4fba0cc387f4fe5