Add npm script for wmf CI to call
BRANCHDEPLOY_AUTH_TOKEN and BRANCHDEPLOY_SITE_ID will be tokens to be
defined in the Jenkins settings.

At some point, it would be nice for the netlify url to be dependent on
something like repo-name + Change-Id so that it stays the same for each
change across its patchsets.

On netlify it will be deployed at the root and so we have to explicitly
change the base path to `/`. Its default for production builds is
otherwise defined in vite.config.js as `/querybuilder/`.

This is intended to work together with I5c23b29111aeeaf5a43b872bf71a10e2a0de20da

Bug: T278706
Change-Id: I37a0b4a82b894671989105230b41ebdf650be3e5