Migration of vue-cli/webpack to vite/rollup
This Involves:
 - removing all vue-cli packages from package.json
 - installing the corrsponding vite plugins
 - Adding backward compatible env variable reader plugin
 - adding vite.js.config file to bundle css, resolve paths and start plugins
 - upgrading node version to 12
 - making sure all the tests, including cypress, pass

Further, build.commonjsOptions.ignoreDynamicRequires = true is needed to
work around an issue with a "require" being misinterpreted as a dynamic
require in sparqljs. That issue resulted in "ReferenceError: "require"
is not defined" in the browser when looking at production builds.

Also, vite does not check for typescript errors before building, so it
is necessary to do that explicitly in the npm build script.

Bug: T276366
Change-Id: I6bd9d9f5ad599454c7b8fa34cf145f69813bcb3e