Do not shadow native Error object
This is a bad practice in general and even though @vue/cli was able to
deal with this, vite is no longer able to handle it and so the cleanest
solution is to use a custom name for this custom interface.

The eslint rule is disabled because type script does type checking
better. See
https://github.com/typescript-eslint/typescript-eslint/blob/75dcf6942286f6ad3b507e469d160dc526cc7274/docs/getting-started/linting/FAQ.md#i-get-errors-from-the-no-undef-rule-about-global-variables-not-being-defined-even-though-there-are-no-typescript-errors

Change-Id: I1c75dd48c0b5b3b929f309de6081b6d138f880c1