Make html title translateable
It is hard coded in the index.html, because that is what is shown before
any javascript is available.

Also, at the moment when we receive the messages, the i18n plugin is not
actually yet available as this.$i18n. Therefore we have to pull out the
right messages manually.

Bug: T285077
Change-Id: I2eafbef32b5ee11fca613cfb7caebc6b636948a0