Refactor the urlShortener into its own repository
This removes the logic for the request to the URL shortener from the
component and into a new repository, where it belongs.

The method in the component could use some more refactoring, but that
can happen later.

The tests didn't work the way they were written. In SharableLink, we're
not using a getter to get the state for serialization but rather
accessing the state directly. This means that overriding a getter in the
test is pointless, because it is never executed.

Change-Id: Iddc69e06a23ce15b4173bc113278eab001f755c4