Load labels after restoring state from URL
After we deserialize and restore the state, send search requests for all
the entity IDs in the query, and (assuming the query hasn’t changed by
the time the searches complete) add those search results to the state,
so that their labels are shown.

This requires a small change in EntityLookup too, so that it updates its
search field (the value shown in the input) when the value updates. For
now, we only do this when the ID of the value didn’t change.

Bug: T280684
Change-Id: I8e42c2b9bf7b5f6b40c36e4bafa5202d06d8107f