Upgrade typescript
This pins the minor version of typescript to the currently newest.
Typescript adds new rules in minor version upgrades, so we should be
intentional about incrementing them.

Some eslint dependencies had to be updated in order for the versions to
not conflict.

Also, some rules changed. `ban-ts-ignore` is now `ban-ts-comment` and a
typescript rule specifically for camelcase no longer exists.

There were more rules breaking, but they have been disabled and will be
tackled in their dedicated commit following this one.

Bug: T311231
Change-Id: I16194695e1aad1f0864323c848f2e2fab7ca4c3f