Update remaining eslint dependencies
Somehow, the newer version of the wikimedia-eslint-config is not pulling
the es-plugin automatically, so we have to do it explicitly.

Also, somehow it now tries to apply all-sorts of ill-fitting rules to
configuration files, so we now ignore them explicitly, which was also
the case until now.

Enabling proper linting for configuration files will be handled in
context of an extra task, as well as making use of the
wikimedia/vue2-common linting configuration for our vuejs files.
See: T311437

Two minor updates to the code were made to satisfy new rules. Also, the
camelcase rule does now for some reason not trigger anymore, so the
suppression is no longer needed.

Bug: T311231
Change-Id: I5df5d9acf47f7dd67a3b692abebc899c97f81122