Fix QueryBuilder <main> block axis padding
According to T288939 (and as implemented in change Ib48f38b07c, commit
f34638d0fe), the block-start (“top”) spacing before the header (which is
part of the <main>, though maybe it shouldn’t be) should always be
$dimension-layout-small; and according to T324356, the block-end
(“bottom”) spacing before the footer (which is outside the <main>)
should always be $dimension-layout-large. Note that the footer still has
some additional spacing of its own at block-start.

All of the <main> spacing is currently specified as padding, so keep it
that way for now, though I personally feel like margin would be more
appropriate. The previous margin-block: auto had been added in commit
c88d44d8be, which was ostensibly only about the inline axis, and as far
as I can tell it had no effect; remove it to avoid confusion.

Bug: T324356
Change-Id: I143a35a1bbbab0e38ae81ec43bfa83f47b47e601