Hide underlines for all links in footer
Including the ones generated by messages, which don’t already have their
underlines hidden like Wikit links have. We put this in a mixin so that
we can later reuse it for the “Show query in the Query Service” link.

Also, add a :focus-visible outline for Chrome; this is just copied from
Wikidata Mismatch Finder. (IMHO the outline is actually quite hard to
see, but if we want to improve it, we should probably do it everywhere.)

Bug: T324356
Change-Id: I2ca1910c175eef9bf07892117d8fd462294cb1d6