Turn ItemValueLookup into EntityValueLookup
ItemValueLookup and LexemeValueLookup were almost identical, the only
difference is one store action and one i18n message. I’m not sure what
the best way to reuse the code is (I was thinking of having the same
file export both components, with an internal helper function to
deduplicate the code, but I don’t think the <template> would work then),
but I think an entityType prop, from which the store action and message
key can be computed, works alright.

Bug: T329204
Change-Id: I286f6f886450805b9f03c85f5aecfb136c65d548