Load language data from @wikimedia/language-data
For now, the languages that can be selected are still just an array of
strings, and the emitted event is an autonym, which isn’t very useful;
we’ll turn this into more structured data (so we can emit a language
code) later.

The languages are also moved into the LanguageSelector, instead of being
a prop; I think this makes the LanguageSelector easier to work with
(especially at the moment, where it’s not actually used yet).

Bug: T329733
Change-Id: I8b258e583c9dd71bba35caab4d6e0660b80c1e85