Do not close LanguageSelector on clicking clear button or language dropdown
It was impossible to use clear button or select a language from dropdown menu because of native blur event triggered by LanguageSelectorInput component.

We can no longer react to the native blur event of the <input>, because that also closes the selector when anything not a child of LanguageSelectorInput component clicked.

Therefore, the LanguageSelectorInput component only emits events on the respective key-presses. Loss of focus due to click/touch events outside is detected separately.

Thanks for help on description: @michael.grosse@wikimedia.de

Bug: T333932
Change-Id: Ie223b9b491b038a061efe9f22476e8eb835cfa9e