Fix image URLs in build
Vite is supposed to prefix asset [1] URLs with the base / origin, and
you can see this working for url() in CSS (otherwise the Wikidata logo
would be broken), but due to vitejs/vite-plugin-vue#85 [2] this doesn’t
work for <img> src= attributes. Work around it by importing the
importing the files and then using that URL dynamically. (In the tests,
we need to map those imports to a fake module so Jest doesn’t try to
load the real file content as JS source.)

[1]: https://vitejs.dev/guide/assets.html
[2]: https://github.com/vitejs/vite-plugin-vue/issues/85

Bug: T335754
Change-Id: Ie3c1cf9c225c646b2f7d94727a634e9ff78aad4e