Fix scrolling being broken on mobile after switching language
In the current implementation, switching the language intentionally
triggers a rerender of basically the entire Query-Builder app in
App.vue. This resets the `showLanguageSelector` data field to its
original value, which means that the language selector is not being
shown. However, that does not trigger the watcher, and thus, the class
from `<body>` is not being removed.

This commit here now makes the watcher "eager", which means it is also
executed when the class is initially rendered. Removing a non-existing
class when initially loading the QueryBuilder is fine, as
DOMTokenList:remove() just does nothing if the class does not exists and
no error is thrown.

See:
* https://vuejs.org/guide/essentials/watchers.html#eager-watchers
* https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList/remove

Bug: T340233
Change-Id: I862988d9ac38e5ff106c81256bc01966170a7877