Query Builder: move scripts from jenkins into repository
The job running these may be found here:
https://gerrit.wikimedia.org/g/integration/config/+/d3fcb3a7a0/jjb/wikidata.yaml#134

Originally the content of those was mostly copied over from the QS GUI
job above it (with the exception of a different build path).

The user name remained 'wdquerybuilder' and email list name is still
'wdqs-gui-build' until we create better named credentials in a separate
task.

Bug: T328543
Change-Id: Ifcb881a77254d07396277af2b29b239a2d228d73