Remove postcss-logical
This 'PostCSS' plugin was installed in order to provide support for
CSS logical properties in older browsers. However, at the time of
this commit, logical properties are supported in the last two
versions of all major browsers that have more than 1% global usage,
matching the requirements specified in `.browserlistrc`.

Removing the post-processor plugin resolves a long-standing issue
with RTL language scripts not appearing properly, due to a breaking
change to a feature we relied on in the plugin itself.

Bug: T371519
Change-Id: I9a67d03c5093ec9768f860c5efb3781678a72ae4