Fix spacing issues caused by non bidi stylesheet
This change updates codex so that we are able to use the new
experimental stylesheet which supports bidi languages.
Additionally, the general codex stylesheet import is moved from a
sub component to the top level component App.vue, where other
general stylesheets are imported.

This change also removes an unnecessary codex stylesheet import in
the stylesheet intended for the noscript message.

Change-Id: Iaf1bb253c479be4989501da775da8ad1d5a01c7b